jws-server/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.23,<1
python-jose[cryptography]>=3.3.0,<4
